ourspace/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.13.3
idna==3.6
oauthlib==3.2.2
phonenumbers==8.13.33
pillow==10.3.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
urllib3==2.2.1
virtualenv==20.25.1
